requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask == 1.1.2
Pillow == 7.0.0
imageio == 2.6.1
numpy == 1.18.4
scikit_image == 0.16.2
~ensorflow_gpu == 2.1.0
